+++ cli/efbs/efbs.cc
@@ -444,6 +444,8 @@ using RawGetter = vlink::Getter<vlink::Bytes>;
 int start_efbs_pub(const std::string& url, const std::string& fbs_dir, const std::string& fbstxt_file,
                    const std::string& fbs_json, const std::string& ser, vlink::SchemaType schema_type,
                    bool use_blob_encoding, bool native_mode, int times, int interval) {
+  const std::string native_ip = native_mode ? vlink::Utils::get_env("VLINK_DDS_NATIVE_IP", "127.0.0.1") : std::string();
+
   if VUNLIKELY (!has_intra_bind && vlink::Url::is_intra_type(url)) {
     std::cerr << "Cannot pub intra url." << std::endl;
     has_quit = true;
@@ -694,7 +696,7 @@ int start_efbs_pub(const std::string& url, const std::string& fbs_dir, const std
   }
 
   if (native_mode) {
-    raw_pub->set_property("dds.ip", "127.0.0.1");
+    raw_pub->set_property("dds.ip", native_ip);
   }
 
   raw_pub->set_ser_type(target_ser, target_schema_type);
@@ -835,6 +837,8 @@ class ParserLoop : public vlink::MessageLoop {
 int start_efbs_sub(const std::string& url, const std::string& fbs_dir, const std::string& ser,
                    vlink::SchemaType schema_type, bool use_blob_encoding, bool native_mode, const std::string& filter,
                    bool use_getter) {
+  const std::string native_ip = native_mode ? vlink::Utils::get_env("VLINK_DDS_NATIVE_IP", "127.0.0.1") : std::string();
+
   if VUNLIKELY (!has_intra_bind && vlink::Url::is_intra_type(url)) {
     std::cerr << "Cannot sub intra url." << std::endl;
     has_quit = true;
@@ -1643,7 +1647,7 @@ int start_efbs_sub(const std::string& url, const std::string& fbs_dir, const std
       raw_getter.emplace(std::make_shared<RawGetter>(url, vlink::InitType::kWithoutInit));
 
       if (native_mode) {
-        (*raw_getter)->set_property("dds.ip", "127.0.0.1");
+        (*raw_getter)->set_property("dds.ip", native_ip);
       }
 
       (*raw_getter)->set_ser_type(target_ser, target_schema_type);
@@ -1653,7 +1657,7 @@ int start_efbs_sub(const std::string& url, const std::string& fbs_dir, const std
       raw_sub.emplace(std::make_shared<RawSub>(url, vlink::InitType::kWithoutInit));
 
       if (native_mode) {
-        (*raw_sub)->set_property("dds.ip", "127.0.0.1");
+        (*raw_sub)->set_property("dds.ip", native_ip);
       }
 
       (*raw_sub)->set_ser_type(target_ser, target_schema_type);